+++ src/main/java/com/thee5176/record/springboot_cqrs_command/Application/controller/HealthCheckController.java
@@ -1,0 +1,14 @@
+package com.thee5176.record.springboot_cqrs_command.Application.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HealthCheckController {
+    @GetMapping("/health")
+    public ResponseEntity<String> healthCheck() {
+        return ResponseEntity.ok("Service is healthy!");
+    }
+
+}
