+++ src/main/java/com/thee5176/record/springboot_cqrs_command/Domain/model/tables/Transactions.java
@@ -1,0 +1,292 @@
+/*
+ * This file is generated by jOOQ.
+ */
+package com.thee5176.record.springboot_cqrs_command.Domain.model.tables;
+
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.util.Collection;
+import java.util.UUID;
+
+import org.jooq.Condition;
+import org.jooq.Field;
+import org.jooq.ForeignKey;
+import org.jooq.InverseForeignKey;
+import org.jooq.Name;
+import org.jooq.Path;
+import org.jooq.PlainSQL;
+import org.jooq.QueryPart;
+import org.jooq.Record;
+import org.jooq.SQL;
+import org.jooq.Schema;
+import org.jooq.Select;
+import org.jooq.Stringly;
+import org.jooq.Table;
+import org.jooq.TableField;
+import org.jooq.TableOptions;
+import org.jooq.UniqueKey;
+import org.jooq.impl.DSL;
+import org.jooq.impl.SQLDataType;
+import org.jooq.impl.TableImpl;
+
+import com.thee5176.record.springboot_cqrs_command.Domain.model.Keys;
+import com.thee5176.record.springboot_cqrs_command.Domain.model.Public;
+import com.thee5176.record.springboot_cqrs_command.Domain.model.tables.Entries.EntriesPath;
+import com.thee5176.record.springboot_cqrs_command.Domain.model.tables.records.TransactionsRecord;
+
+
+/**
+ * This class is generated by jOOQ.
+ */
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
+public class Transactions extends TableImpl<TransactionsRecord> {
+
+    private static final long serialVersionUID = 1L;
+
+    /**
+     * The reference instance of <code>public.transactions</code>
+     */
+    public static final Transactions TRANSACTIONS = new Transactions();
+
+    /**
+     * The class holding records for this type
+     */
+    @Override
+    public Class<TransactionsRecord> getRecordType() {
+        return TransactionsRecord.class;
+    }
+
+    /**
+     * The column <code>public.transactions.id</code>.
+     */
+    public final TableField<TransactionsRecord, UUID> ID = createField(DSL.name("id"), SQLDataType.UUID.nullable(false), this, "");
+
+    /**
+     * The column <code>public.transactions.date</code>.
+     */
+    public final TableField<TransactionsRecord, LocalDate> DATE = createField(DSL.name("date"), SQLDataType.LOCALDATE.nullable(false), this, "");
+
+    /**
+     * The column <code>public.transactions.description</code>.
+     */
+    public final TableField<TransactionsRecord, String> DESCRIPTION = createField(DSL.name("description"), SQLDataType.VARCHAR.nullable(false), this, "");
+
+    /**
+     * The column <code>public.transactions.created_at</code>.
+     */
+    public final TableField<TransactionsRecord, LocalDateTime> CREATED_AT = createField(DSL.name("created_at"), SQLDataType.LOCALDATETIME(6).nullable(false), this, "");
+
+    /**
+     * The column <code>public.transactions.updated_at</code>.
+     */
+    public final TableField<TransactionsRecord, LocalDateTime> UPDATED_AT = createField(DSL.name("updated_at"), SQLDataType.LOCALDATETIME(6).nullable(false), this, "");
+
+    private Transactions(Name alias, Table<TransactionsRecord> aliased) {
+        this(alias, aliased, (Field<?>[]) null, null);
+    }
+
+    private Transactions(Name alias, Table<TransactionsRecord> aliased, Field<?>[] parameters, Condition where) {
+        super(alias, null, aliased, parameters, DSL.comment(""), TableOptions.table(), where);
+    }
+
+    /**
+     * Create an aliased <code>public.transactions</code> table reference
+     */
+    public Transactions(String alias) {
+        this(DSL.name(alias), TRANSACTIONS);
+    }
+
+    /**
+     * Create an aliased <code>public.transactions</code> table reference
+     */
+    public Transactions(Name alias) {
+        this(alias, TRANSACTIONS);
+    }
+
+    /**
+     * Create a <code>public.transactions</code> table reference
+     */
+    public Transactions() {
+        this(DSL.name("transactions"), null);
+    }
+
+    public <O extends Record> Transactions(Table<O> path, ForeignKey<O, TransactionsRecord> childPath, InverseForeignKey<O, TransactionsRecord> parentPath) {
+        super(path, childPath, parentPath, TRANSACTIONS);
+    }
+
+    /**
+     * A subtype implementing {@link Path} for simplified path-based joins.
+     */
+    public static class TransactionsPath extends Transactions implements Path<TransactionsRecord> {
+
+        private static final long serialVersionUID = 1L;
+        public <O extends Record> TransactionsPath(Table<O> path, ForeignKey<O, TransactionsRecord> childPath, InverseForeignKey<O, TransactionsRecord> parentPath) {
+            super(path, childPath, parentPath);
+        }
+        private TransactionsPath(Name alias, Table<TransactionsRecord> aliased) {
+            super(alias, aliased);
+        }
+
+        @Override
+        public TransactionsPath as(String alias) {
+            return new TransactionsPath(DSL.name(alias), this);
+        }
+
+        @Override
+        public TransactionsPath as(Name alias) {
+            return new TransactionsPath(alias, this);
+        }
+
+        @Override
+        public TransactionsPath as(Table<?> alias) {
+            return new TransactionsPath(alias.getQualifiedName(), this);
+        }
+    }
+
+    @Override
+    public Schema getSchema() {
+        return aliased() ? null : Public.PUBLIC;
+    }
+
+    @Override
+    public UniqueKey<TransactionsRecord> getPrimaryKey() {
+        return Keys.TRANSACTIONS_PKEY;
+    }
+
+    private transient EntriesPath _entries;
+
+    /**
+     * Get the implicit to-many join path to the <code>public.entries</code>
+     * table
+     */
+    public EntriesPath entries() {
+        if (_entries == null)
+            _entries = new EntriesPath(this, null, Keys.ENTRIES__ENTRIES_TRANSACTION_ID_FKEY.getInverseKey());
+
+        return _entries;
+    }
+
+    @Override
+    public Transactions as(String alias) {
+        return new Transactions(DSL.name(alias), this);
+    }
+
+    @Override
+    public Transactions as(Name alias) {
+        return new Transactions(alias, this);
+    }
+
+    @Override
+    public Transactions as(Table<?> alias) {
+        return new Transactions(alias.getQualifiedName(), this);
+    }
+
+    /**
+     * Rename this table
+     */
+    @Override
+    public Transactions rename(String name) {
+        return new Transactions(DSL.name(name), null);
+    }
+
+    /**
+     * Rename this table
+     */
+    @Override
+    public Transactions rename(Name name) {
+        return new Transactions(name, null);
+    }
+
+    /**
+     * Rename this table
+     */
+    @Override
+    public Transactions rename(Table<?> name) {
+        return new Transactions(name.getQualifiedName(), null);
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    public Transactions where(Condition condition) {
+        return new Transactions(getQualifiedName(), aliased() ? this : null, null, condition);
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    public Transactions where(Collection<? extends Condition> conditions) {
+        return where(DSL.and(conditions));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    public Transactions where(Condition... conditions) {
+        return where(DSL.and(conditions));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    public Transactions where(Field<Boolean> condition) {
+        return where(DSL.condition(condition));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    @PlainSQL
+    public Transactions where(SQL condition) {
+        return where(DSL.condition(condition));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    @PlainSQL
+    public Transactions where(@Stringly.SQL String condition) {
+        return where(DSL.condition(condition));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    @PlainSQL
+    public Transactions where(@Stringly.SQL String condition, Object... binds) {
+        return where(DSL.condition(condition, binds));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    @PlainSQL
+    public Transactions where(@Stringly.SQL String condition, QueryPart... parts) {
+        return where(DSL.condition(condition, parts));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    public Transactions whereExists(Select<?> select) {
+        return where(DSL.exists(select));
+    }
+
+    /**
+     * Create an inline derived table from this table
+     */
+    @Override
+    public Transactions whereNotExists(Select<?> select) {
+        return where(DSL.notExists(select));
+    }
+}
