+++ src/main/java/com/thee5176/record/springboot_cqrs_command/Domain/model/tables/pojos/FlywaySchemaHistory.java
@@ -1,0 +1,326 @@
+/*
+ * This file is generated by jOOQ.
+ */
+package com.thee5176.record.springboot_cqrs_command.Domain.model.tables.pojos;
+
+
+import java.io.Serializable;
+import java.time.LocalDateTime;
+
+
+/**
+ * This class is generated by jOOQ.
+ */
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
+public class FlywaySchemaHistory implements Serializable {
+
+    private static final long serialVersionUID = 1L;
+
+    private Integer installedRank;
+    private String version;
+    private String description;
+    private String type;
+    private String script;
+    private Integer checksum;
+    private String installedBy;
+    private LocalDateTime installedOn;
+    private Integer executionTime;
+    private Boolean success;
+
+    public FlywaySchemaHistory() {}
+
+    public FlywaySchemaHistory(FlywaySchemaHistory value) {
+        this.installedRank = value.installedRank;
+        this.version = value.version;
+        this.description = value.description;
+        this.type = value.type;
+        this.script = value.script;
+        this.checksum = value.checksum;
+        this.installedBy = value.installedBy;
+        this.installedOn = value.installedOn;
+        this.executionTime = value.executionTime;
+        this.success = value.success;
+    }
+
+    public FlywaySchemaHistory(
+        Integer installedRank,
+        String version,
+        String description,
+        String type,
+        String script,
+        Integer checksum,
+        String installedBy,
+        LocalDateTime installedOn,
+        Integer executionTime,
+        Boolean success
+    ) {
+        this.installedRank = installedRank;
+        this.version = version;
+        this.description = description;
+        this.type = type;
+        this.script = script;
+        this.checksum = checksum;
+        this.installedBy = installedBy;
+        this.installedOn = installedOn;
+        this.executionTime = executionTime;
+        this.success = success;
+    }
+
+    /**
+     * Getter for <code>public.flyway_schema_history.installed_rank</code>.
+     */
+    public Integer getInstalledRank() {
+        return this.installedRank;
+    }
+
+    /**
+     * Setter for <code>public.flyway_schema_history.installed_rank</code>.
+     */
+    public FlywaySchemaHistory setInstalledRank(Integer installedRank) {
+        this.installedRank = installedRank;
+        return this;
+    }
+
+    /**
+     * Getter for <code>public.flyway_schema_history.version</code>.
+     */
+    public String getVersion() {
+        return this.version;
+    }
+
+    /**
+     * Setter for <code>public.flyway_schema_history.version</code>.
+     */
+    public FlywaySchemaHistory setVersion(String version) {
+        this.version = version;
+        return this;
+    }
+
+    /**
+     * Getter for <code>public.flyway_schema_history.description</code>.
+     */
+    public String getDescription() {
+        return this.description;
+    }
+
+    /**
+     * Setter for <code>public.flyway_schema_history.description</code>.
+     */
+    public FlywaySchemaHistory setDescription(String description) {
+        this.description = description;
+        return this;
+    }
+
+    /**
+     * Getter for <code>public.flyway_schema_history.type</code>.
+     */
+    public String getType() {
+        return this.type;
+    }
+
+    /**
+     * Setter for <code>public.flyway_schema_history.type</code>.
+     */
+    public FlywaySchemaHistory setType(String type) {
+        this.type = type;
+        return this;
+    }
+
+    /**
+     * Getter for <code>public.flyway_schema_history.script</code>.
+     */
+    public String getScript() {
+        return this.script;
+    }
+
+    /**
+     * Setter for <code>public.flyway_schema_history.script</code>.
+     */
+    public FlywaySchemaHistory setScript(String script) {
+        this.script = script;
+        return this;
+    }
+
+    /**
+     * Getter for <code>public.flyway_schema_history.checksum</code>.
+     */
+    public Integer getChecksum() {
+        return this.checksum;
+    }
+
+    /**
+     * Setter for <code>public.flyway_schema_history.checksum</code>.
+     */
+    public FlywaySchemaHistory setChecksum(Integer checksum) {
+        this.checksum = checksum;
+        return this;
+    }
+
+    /**
+     * Getter for <code>public.flyway_schema_history.installed_by</code>.
+     */
+    public String getInstalledBy() {
+        return this.installedBy;
+    }
+
+    /**
+     * Setter for <code>public.flyway_schema_history.installed_by</code>.
+     */
+    public FlywaySchemaHistory setInstalledBy(String installedBy) {
+        this.installedBy = installedBy;
+        return this;
+    }
+
+    /**
+     * Getter for <code>public.flyway_schema_history.installed_on</code>.
+     */
+    public LocalDateTime getInstalledOn() {
+        return this.installedOn;
+    }
+
+    /**
+     * Setter for <code>public.flyway_schema_history.installed_on</code>.
+     */
+    public FlywaySchemaHistory setInstalledOn(LocalDateTime installedOn) {
+        this.installedOn = installedOn;
+        return this;
+    }
+
+    /**
+     * Getter for <code>public.flyway_schema_history.execution_time</code>.
+     */
+    public Integer getExecutionTime() {
+        return this.executionTime;
+    }
+
+    /**
+     * Setter for <code>public.flyway_schema_history.execution_time</code>.
+     */
+    public FlywaySchemaHistory setExecutionTime(Integer executionTime) {
+        this.executionTime = executionTime;
+        return this;
+    }
+
+    /**
+     * Getter for <code>public.flyway_schema_history.success</code>.
+     */
+    public Boolean getSuccess() {
+        return this.success;
+    }
+
+    /**
+     * Setter for <code>public.flyway_schema_history.success</code>.
+     */
+    public FlywaySchemaHistory setSuccess(Boolean success) {
+        this.success = success;
+        return this;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj)
+            return true;
+        if (obj == null)
+            return false;
+        if (getClass() != obj.getClass())
+            return false;
+        final FlywaySchemaHistory other = (FlywaySchemaHistory) obj;
+        if (this.installedRank == null) {
+            if (other.installedRank != null)
+                return false;
+        }
+        else if (!this.installedRank.equals(other.installedRank))
+            return false;
+        if (this.version == null) {
+            if (other.version != null)
+                return false;
+        }
+        else if (!this.version.equals(other.version))
+            return false;
+        if (this.description == null) {
+            if (other.description != null)
+                return false;
+        }
+        else if (!this.description.equals(other.description))
+            return false;
+        if (this.type == null) {
+            if (other.type != null)
+                return false;
+        }
+        else if (!this.type.equals(other.type))
+            return false;
+        if (this.script == null) {
+            if (other.script != null)
+                return false;
+        }
+        else if (!this.script.equals(other.script))
+            return false;
+        if (this.checksum == null) {
+            if (other.checksum != null)
+                return false;
+        }
+        else if (!this.checksum.equals(other.checksum))
+            return false;
+        if (this.installedBy == null) {
+            if (other.installedBy != null)
+                return false;
+        }
+        else if (!this.installedBy.equals(other.installedBy))
+            return false;
+        if (this.installedOn == null) {
+            if (other.installedOn != null)
+                return false;
+        }
+        else if (!this.installedOn.equals(other.installedOn))
+            return false;
+        if (this.executionTime == null) {
+            if (other.executionTime != null)
+                return false;
+        }
+        else if (!this.executionTime.equals(other.executionTime))
+            return false;
+        if (this.success == null) {
+            if (other.success != null)
+                return false;
+        }
+        else if (!this.success.equals(other.success))
+            return false;
+        return true;
+    }
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + ((this.installedRank == null) ? 0 : this.installedRank.hashCode());
+        result = prime * result + ((this.version == null) ? 0 : this.version.hashCode());
+        result = prime * result + ((this.description == null) ? 0 : this.description.hashCode());
+        result = prime * result + ((this.type == null) ? 0 : this.type.hashCode());
+        result = prime * result + ((this.script == null) ? 0 : this.script.hashCode());
+        result = prime * result + ((this.checksum == null) ? 0 : this.checksum.hashCode());
+        result = prime * result + ((this.installedBy == null) ? 0 : this.installedBy.hashCode());
+        result = prime * result + ((this.installedOn == null) ? 0 : this.installedOn.hashCode());
+        result = prime * result + ((this.executionTime == null) ? 0 : this.executionTime.hashCode());
+        result = prime * result + ((this.success == null) ? 0 : this.success.hashCode());
+        return result;
+    }
+
+    @Override
+    public String toString() {
+        StringBuilder sb = new StringBuilder("FlywaySchemaHistory (");
+
+        sb.append(installedRank);
+        sb.append(", ").append(version);
+        sb.append(", ").append(description);
+        sb.append(", ").append(type);
+        sb.append(", ").append(script);
+        sb.append(", ").append(checksum);
+        sb.append(", ").append(installedBy);
+        sb.append(", ").append(installedOn);
+        sb.append(", ").append(executionTime);
+        sb.append(", ").append(success);
+
+        sb.append(")");
+        return sb.toString();
+    }
+}
