+++ src/main/java/com/thee5176/record/springboot_cqrs_command/Domain/model/tables/records/EntriesRecord.java
@@ -1,0 +1,183 @@
+/*
+ * This file is generated by jOOQ.
+ */
+package com.thee5176.record.springboot_cqrs_command.Domain.model.tables.records;
+
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+import org.jooq.Record1;
+import org.jooq.impl.UpdatableRecordImpl;
+
+import com.thee5176.record.springboot_cqrs_command.Domain.model.enums.BalanceType;
+import com.thee5176.record.springboot_cqrs_command.Domain.model.tables.Entries;
+
+
+/**
+ * This class is generated by jOOQ.
+ */
+@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
+public class EntriesRecord extends UpdatableRecordImpl<EntriesRecord> {
+
+    private static final long serialVersionUID = 1L;
+
+    /**
+     * Setter for <code>public.entries.id</code>.
+     */
+    public EntriesRecord setId(UUID value) {
+        set(0, value);
+        return this;
+    }
+
+    /**
+     * Getter for <code>public.entries.id</code>.
+     */
+    public UUID getId() {
+        return (UUID) get(0);
+    }
+
+    /**
+     * Setter for <code>public.entries.transaction_id</code>.
+     */
+    public EntriesRecord setTransactionId(UUID value) {
+        set(1, value);
+        return this;
+    }
+
+    /**
+     * Getter for <code>public.entries.transaction_id</code>.
+     */
+    public UUID getTransactionId() {
+        return (UUID) get(1);
+    }
+
+    /**
+     * Setter for <code>public.entries.coa</code>.
+     */
+    public EntriesRecord setCoa(Integer value) {
+        set(2, value);
+        return this;
+    }
+
+    /**
+     * Getter for <code>public.entries.coa</code>.
+     */
+    public Integer getCoa() {
+        return (Integer) get(2);
+    }
+
+    /**
+     * Setter for <code>public.entries.amount</code>.
+     */
+    public EntriesRecord setAmount(Double value) {
+        set(3, value);
+        return this;
+    }
+
+    /**
+     * Getter for <code>public.entries.amount</code>.
+     */
+    public Double getAmount() {
+        return (Double) get(3);
+    }
+
+    /**
+     * Setter for <code>public.entries.type</code>.
+     */
+    public EntriesRecord setType(BalanceType value) {
+        set(4, value);
+        return this;
+    }
+
+    /**
+     * Getter for <code>public.entries.type</code>.
+     */
+    public BalanceType getType() {
+        return (BalanceType) get(4);
+    }
+
+    /**
+     * Setter for <code>public.entries.created_at</code>.
+     */
+    public EntriesRecord setCreatedAt(LocalDateTime value) {
+        set(5, value);
+        return this;
+    }
+
+    /**
+     * Getter for <code>public.entries.created_at</code>.
+     */
+    public LocalDateTime getCreatedAt() {
+        return (LocalDateTime) get(5);
+    }
+
+    /**
+     * Setter for <code>public.entries.updated_at</code>.
+     */
+    public EntriesRecord setUpdatedAt(LocalDateTime value) {
+        set(6, value);
+        return this;
+    }
+
+    /**
+     * Getter for <code>public.entries.updated_at</code>.
+     */
+    public LocalDateTime getUpdatedAt() {
+        return (LocalDateTime) get(6);
+    }
+
+    // -------------------------------------------------------------------------
+    // Primary key information
+    // -------------------------------------------------------------------------
+
+    @Override
+    public Record1<UUID> key() {
+        return (Record1) super.key();
+    }
+
+    // -------------------------------------------------------------------------
+    // Constructors
+    // -------------------------------------------------------------------------
+
+    /**
+     * Create a detached EntriesRecord
+     */
+    public EntriesRecord() {
+        super(Entries.ENTRIES);
+    }
+
+    /**
+     * Create a detached, initialised EntriesRecord
+     */
+    public EntriesRecord(UUID id, UUID transactionId, Integer coa, Double amount, BalanceType type, LocalDateTime createdAt, LocalDateTime updatedAt) {
+        super(Entries.ENTRIES);
+
+        setId(id);
+        setTransactionId(transactionId);
+        setCoa(coa);
+        setAmount(amount);
+        setType(type);
+        setCreatedAt(createdAt);
+        setUpdatedAt(updatedAt);
+        resetChangedOnNotNull();
+    }
+
+    /**
+     * Create a detached, initialised EntriesRecord
+     */
+    public EntriesRecord(com.thee5176.record.springboot_cqrs_command.Domain.model.tables.pojos.Entries value) {
+        super(Entries.ENTRIES);
+
+        if (value != null) {
+            setId(value.getId());
+            setTransactionId(value.getTransactionId());
+            setCoa(value.getCoa());
+            setAmount(value.getAmount());
+            setType(value.getType());
+            setCreatedAt(value.getCreatedAt());
+            setUpdatedAt(value.getUpdatedAt());
+            resetChangedOnNotNull();
+        }
+    }
+}
