--- src/main/java/com/thee5176/record/springboot_cqrs_command/model/Public.java
@@ -1,75 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.thee5176.record.springboot_cqrs_command.model;
-
-
-import com.thee5176.record.springboot_cqrs_command.model.tables.CodeOfAccount;
-import com.thee5176.record.springboot_cqrs_command.model.tables.Entries;
-import com.thee5176.record.springboot_cqrs_command.model.tables.FlywaySchemaHistory;
-import com.thee5176.record.springboot_cqrs_command.model.tables.Transactions;
-
-import java.util.Arrays;
-import java.util.List;
-
-import org.jooq.Catalog;
-import org.jooq.Table;
-import org.jooq.impl.SchemaImpl;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
-public class Public extends SchemaImpl {
-
-    private static final long serialVersionUID = 1L;
-
-    /**
-     * The reference instance of <code>public</code>
-     */
-    public static final Public PUBLIC = new Public();
-
-    /**
-     * The table <code>public.code_of_account</code>.
-     */
-    public final CodeOfAccount CODE_OF_ACCOUNT = CodeOfAccount.CODE_OF_ACCOUNT;
-
-    /**
-     * The table <code>public.entries</code>.
-     */
-    public final Entries ENTRIES = Entries.ENTRIES;
-
-    /**
-     * The table <code>public.flyway_schema_history</code>.
-     */
-    public final FlywaySchemaHistory FLYWAY_SCHEMA_HISTORY = FlywaySchemaHistory.FLYWAY_SCHEMA_HISTORY;
-
-    /**
-     * The table <code>public.transactions</code>.
-     */
-    public final Transactions TRANSACTIONS = Transactions.TRANSACTIONS;
-
-    /**
-     * No further instances allowed
-     */
-    private Public() {
-        super("public", null);
-    }
-
-
-    @Override
-    public Catalog getCatalog() {
-        return DefaultCatalog.DEFAULT_CATALOG;
-    }
-
-    @Override
-    public final List<Table<?>> getTables() {
-        return Arrays.asList(
-            CodeOfAccount.CODE_OF_ACCOUNT,
-            Entries.ENTRIES,
-            FlywaySchemaHistory.FLYWAY_SCHEMA_HISTORY,
-            Transactions.TRANSACTIONS
-        );
-    }
-}
