--- src/main/java/com/thee5176/record/springboot_cqrs_command/model/Tables.java
@@ -1,38 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.thee5176.record.springboot_cqrs_command.model;
-
-
-import com.thee5176.record.springboot_cqrs_command.model.tables.CodeOfAccount;
-import com.thee5176.record.springboot_cqrs_command.model.tables.Entries;
-import com.thee5176.record.springboot_cqrs_command.model.tables.FlywaySchemaHistory;
-import com.thee5176.record.springboot_cqrs_command.model.tables.Transactions;
-
-
-/**
- * Convenience access to all tables in public.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
-public class Tables {
-
-    /**
-     * The table <code>public.code_of_account</code>.
-     */
-    public static final CodeOfAccount CODE_OF_ACCOUNT = CodeOfAccount.CODE_OF_ACCOUNT;
-
-    /**
-     * The table <code>public.entries</code>.
-     */
-    public static final Entries ENTRIES = Entries.ENTRIES;
-
-    /**
-     * The table <code>public.flyway_schema_history</code>.
-     */
-    public static final FlywaySchemaHistory FLYWAY_SCHEMA_HISTORY = FlywaySchemaHistory.FLYWAY_SCHEMA_HISTORY;
-
-    /**
-     * The table <code>public.transactions</code>.
-     */
-    public static final Transactions TRANSACTIONS = Transactions.TRANSACTIONS;
-}
