--- src/main/java/com/thee5176/record/springboot_cqrs_command/model/enums/Element.java
@@ -1,66 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.thee5176.record.springboot_cqrs_command.model.enums;
-
-
-import com.thee5176.record.springboot_cqrs_command.model.Public;
-
-import org.jooq.Catalog;
-import org.jooq.EnumType;
-import org.jooq.Schema;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
-public enum Element implements EnumType {
-
-    Assets("Assets"),
-
-    Liabilities("Liabilities"),
-
-    Equity("Equity"),
-
-    Revenue("Revenue"),
-
-    Expenses("Expenses"),
-
-    Other("Other");
-
-    private final String literal;
-
-    private Element(String literal) {
-        this.literal = literal;
-    }
-
-    @Override
-    public Catalog getCatalog() {
-        return getSchema().getCatalog();
-    }
-
-    @Override
-    public Schema getSchema() {
-        return Public.PUBLIC;
-    }
-
-    @Override
-    public String getName() {
-        return "element";
-    }
-
-    @Override
-    public String getLiteral() {
-        return literal;
-    }
-
-    /**
-     * Lookup a value of this EnumType by its literal. Returns
-     * <code>null</code>, if no such value could be found, see {@link
-     * EnumType#lookupLiteral(Class, String)}.
-     */
-    public static Element lookupLiteral(String literal) {
-        return EnumType.lookupLiteral(Element.class, literal);
-    }
-}
