--- src/main/java/com/thee5176/record/springboot_cqrs_command/model/tables/Entries.java
@@ -1,322 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.thee5176.record.springboot_cqrs_command.model.tables;
-
-
-import com.thee5176.record.springboot_cqrs_command.model.Keys;
-import com.thee5176.record.springboot_cqrs_command.model.Public;
-import com.thee5176.record.springboot_cqrs_command.model.enums.BalanceType;
-import com.thee5176.record.springboot_cqrs_command.model.tables.CodeOfAccount.CodeOfAccountPath;
-import com.thee5176.record.springboot_cqrs_command.model.tables.Transactions.TransactionsPath;
-import com.thee5176.record.springboot_cqrs_command.model.tables.records.EntriesRecord;
-
-import java.time.LocalDateTime;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.List;
-import java.util.UUID;
-
-import org.jooq.Condition;
-import org.jooq.Field;
-import org.jooq.ForeignKey;
-import org.jooq.InverseForeignKey;
-import org.jooq.Name;
-import org.jooq.Path;
-import org.jooq.PlainSQL;
-import org.jooq.QueryPart;
-import org.jooq.Record;
-import org.jooq.SQL;
-import org.jooq.Schema;
-import org.jooq.Select;
-import org.jooq.Stringly;
-import org.jooq.Table;
-import org.jooq.TableField;
-import org.jooq.TableOptions;
-import org.jooq.UniqueKey;
-import org.jooq.impl.DSL;
-import org.jooq.impl.SQLDataType;
-import org.jooq.impl.TableImpl;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
-public class Entries extends TableImpl<EntriesRecord> {
-
-    private static final long serialVersionUID = 1L;
-
-    /**
-     * The reference instance of <code>public.entries</code>
-     */
-    public static final Entries ENTRIES = new Entries();
-
-    /**
-     * The class holding records for this type
-     */
-    @Override
-    public Class<EntriesRecord> getRecordType() {
-        return EntriesRecord.class;
-    }
-
-    /**
-     * The column <code>public.entries.id</code>.
-     */
-    public final TableField<EntriesRecord, UUID> ID = createField(DSL.name("id"), SQLDataType.UUID.nullable(false), this, "");
-
-    /**
-     * The column <code>public.entries.transaction_id</code>.
-     */
-    public final TableField<EntriesRecord, UUID> TRANSACTION_ID = createField(DSL.name("transaction_id"), SQLDataType.UUID.nullable(false), this, "");
-
-    /**
-     * The column <code>public.entries.coa</code>.
-     */
-    public final TableField<EntriesRecord, Integer> COA = createField(DSL.name("coa"), SQLDataType.INTEGER.nullable(false), this, "");
-
-    /**
-     * The column <code>public.entries.amount</code>.
-     */
-    public final TableField<EntriesRecord, Double> AMOUNT = createField(DSL.name("amount"), SQLDataType.DOUBLE.nullable(false), this, "");
-
-    /**
-     * The column <code>public.entries.type</code>.
-     */
-    public final TableField<EntriesRecord, BalanceType> TYPE = createField(DSL.name("type"), SQLDataType.VARCHAR.nullable(false).asEnumDataType(BalanceType.class), this, "");
-
-    /**
-     * The column <code>public.entries.created_at</code>.
-     */
-    public final TableField<EntriesRecord, LocalDateTime> CREATED_AT = createField(DSL.name("created_at"), SQLDataType.LOCALDATETIME(6).nullable(false), this, "");
-
-    /**
-     * The column <code>public.entries.updated_at</code>.
-     */
-    public final TableField<EntriesRecord, LocalDateTime> UPDATED_AT = createField(DSL.name("updated_at"), SQLDataType.LOCALDATETIME(6).nullable(false), this, "");
-
-    private Entries(Name alias, Table<EntriesRecord> aliased) {
-        this(alias, aliased, (Field<?>[]) null, null);
-    }
-
-    private Entries(Name alias, Table<EntriesRecord> aliased, Field<?>[] parameters, Condition where) {
-        super(alias, null, aliased, parameters, DSL.comment(""), TableOptions.table(), where);
-    }
-
-    /**
-     * Create an aliased <code>public.entries</code> table reference
-     */
-    public Entries(String alias) {
-        this(DSL.name(alias), ENTRIES);
-    }
-
-    /**
-     * Create an aliased <code>public.entries</code> table reference
-     */
-    public Entries(Name alias) {
-        this(alias, ENTRIES);
-    }
-
-    /**
-     * Create a <code>public.entries</code> table reference
-     */
-    public Entries() {
-        this(DSL.name("entries"), null);
-    }
-
-    public <O extends Record> Entries(Table<O> path, ForeignKey<O, EntriesRecord> childPath, InverseForeignKey<O, EntriesRecord> parentPath) {
-        super(path, childPath, parentPath, ENTRIES);
-    }
-
-    /**
-     * A subtype implementing {@link Path} for simplified path-based joins.
-     */
-    public static class EntriesPath extends Entries implements Path<EntriesRecord> {
-
-        private static final long serialVersionUID = 1L;
-        public <O extends Record> EntriesPath(Table<O> path, ForeignKey<O, EntriesRecord> childPath, InverseForeignKey<O, EntriesRecord> parentPath) {
-            super(path, childPath, parentPath);
-        }
-        private EntriesPath(Name alias, Table<EntriesRecord> aliased) {
-            super(alias, aliased);
-        }
-
-        @Override
-        public EntriesPath as(String alias) {
-            return new EntriesPath(DSL.name(alias), this);
-        }
-
-        @Override
-        public EntriesPath as(Name alias) {
-            return new EntriesPath(alias, this);
-        }
-
-        @Override
-        public EntriesPath as(Table<?> alias) {
-            return new EntriesPath(alias.getQualifiedName(), this);
-        }
-    }
-
-    @Override
-    public Schema getSchema() {
-        return aliased() ? null : Public.PUBLIC;
-    }
-
-    @Override
-    public UniqueKey<EntriesRecord> getPrimaryKey() {
-        return Keys.ENTRIES_PKEY;
-    }
-
-    @Override
-    public List<ForeignKey<EntriesRecord, ?>> getReferences() {
-        return Arrays.asList(Keys.ENTRIES__ENTRIES_COA_FKEY, Keys.ENTRIES__ENTRIES_TRANSACTION_ID_FKEY);
-    }
-
-    private transient CodeOfAccountPath _codeOfAccount;
-
-    /**
-     * Get the implicit join path to the <code>public.code_of_account</code>
-     * table.
-     */
-    public CodeOfAccountPath codeOfAccount() {
-        if (_codeOfAccount == null)
-            _codeOfAccount = new CodeOfAccountPath(this, Keys.ENTRIES__ENTRIES_COA_FKEY, null);
-
-        return _codeOfAccount;
-    }
-
-    private transient TransactionsPath _transactions;
-
-    /**
-     * Get the implicit join path to the <code>public.transactions</code> table.
-     */
-    public TransactionsPath transactions() {
-        if (_transactions == null)
-            _transactions = new TransactionsPath(this, Keys.ENTRIES__ENTRIES_TRANSACTION_ID_FKEY, null);
-
-        return _transactions;
-    }
-
-    @Override
-    public Entries as(String alias) {
-        return new Entries(DSL.name(alias), this);
-    }
-
-    @Override
-    public Entries as(Name alias) {
-        return new Entries(alias, this);
-    }
-
-    @Override
-    public Entries as(Table<?> alias) {
-        return new Entries(alias.getQualifiedName(), this);
-    }
-
-    /**
-     * Rename this table
-     */
-    @Override
-    public Entries rename(String name) {
-        return new Entries(DSL.name(name), null);
-    }
-
-    /**
-     * Rename this table
-     */
-    @Override
-    public Entries rename(Name name) {
-        return new Entries(name, null);
-    }
-
-    /**
-     * Rename this table
-     */
-    @Override
-    public Entries rename(Table<?> name) {
-        return new Entries(name.getQualifiedName(), null);
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    public Entries where(Condition condition) {
-        return new Entries(getQualifiedName(), aliased() ? this : null, null, condition);
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    public Entries where(Collection<? extends Condition> conditions) {
-        return where(DSL.and(conditions));
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    public Entries where(Condition... conditions) {
-        return where(DSL.and(conditions));
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    public Entries where(Field<Boolean> condition) {
-        return where(DSL.condition(condition));
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    @PlainSQL
-    public Entries where(SQL condition) {
-        return where(DSL.condition(condition));
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    @PlainSQL
-    public Entries where(@Stringly.SQL String condition) {
-        return where(DSL.condition(condition));
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    @PlainSQL
-    public Entries where(@Stringly.SQL String condition, Object... binds) {
-        return where(DSL.condition(condition, binds));
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    @PlainSQL
-    public Entries where(@Stringly.SQL String condition, QueryPart... parts) {
-        return where(DSL.condition(condition, parts));
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    public Entries whereExists(Select<?> select) {
-        return where(DSL.exists(select));
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    public Entries whereNotExists(Select<?> select) {
-        return where(DSL.notExists(select));
-    }
-}
