--- src/main/java/com/thee5176/record/springboot_cqrs_command/model/tables/pojos/CodeOfAccount.java
@@ -1,193 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.thee5176.record.springboot_cqrs_command.model.tables.pojos;
-
-
-import com.thee5176.record.springboot_cqrs_command.model.enums.BalanceType;
-import com.thee5176.record.springboot_cqrs_command.model.enums.Element;
-
-import java.io.Serializable;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
-public class CodeOfAccount implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-    private Integer code;
-    private String title;
-    private Integer level;
-    private Element element;
-    private BalanceType type;
-
-    public CodeOfAccount() {}
-
-    public CodeOfAccount(CodeOfAccount value) {
-        this.code = value.code;
-        this.title = value.title;
-        this.level = value.level;
-        this.element = value.element;
-        this.type = value.type;
-    }
-
-    public CodeOfAccount(
-        Integer code,
-        String title,
-        Integer level,
-        Element element,
-        BalanceType type
-    ) {
-        this.code = code;
-        this.title = title;
-        this.level = level;
-        this.element = element;
-        this.type = type;
-    }
-
-    /**
-     * Getter for <code>public.code_of_account.code</code>.
-     */
-    public Integer getCode() {
-        return this.code;
-    }
-
-    /**
-     * Setter for <code>public.code_of_account.code</code>.
-     */
-    public CodeOfAccount setCode(Integer code) {
-        this.code = code;
-        return this;
-    }
-
-    /**
-     * Getter for <code>public.code_of_account.title</code>.
-     */
-    public String getTitle() {
-        return this.title;
-    }
-
-    /**
-     * Setter for <code>public.code_of_account.title</code>.
-     */
-    public CodeOfAccount setTitle(String title) {
-        this.title = title;
-        return this;
-    }
-
-    /**
-     * Getter for <code>public.code_of_account.level</code>.
-     */
-    public Integer getLevel() {
-        return this.level;
-    }
-
-    /**
-     * Setter for <code>public.code_of_account.level</code>.
-     */
-    public CodeOfAccount setLevel(Integer level) {
-        this.level = level;
-        return this;
-    }
-
-    /**
-     * Getter for <code>public.code_of_account.element</code>.
-     */
-    public Element getElement() {
-        return this.element;
-    }
-
-    /**
-     * Setter for <code>public.code_of_account.element</code>.
-     */
-    public CodeOfAccount setElement(Element element) {
-        this.element = element;
-        return this;
-    }
-
-    /**
-     * Getter for <code>public.code_of_account.type</code>.
-     */
-    public BalanceType getType() {
-        return this.type;
-    }
-
-    /**
-     * Setter for <code>public.code_of_account.type</code>.
-     */
-    public CodeOfAccount setType(BalanceType type) {
-        this.type = type;
-        return this;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (obj == null)
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        final CodeOfAccount other = (CodeOfAccount) obj;
-        if (this.code == null) {
-            if (other.code != null)
-                return false;
-        }
-        else if (!this.code.equals(other.code))
-            return false;
-        if (this.title == null) {
-            if (other.title != null)
-                return false;
-        }
-        else if (!this.title.equals(other.title))
-            return false;
-        if (this.level == null) {
-            if (other.level != null)
-                return false;
-        }
-        else if (!this.level.equals(other.level))
-            return false;
-        if (this.element == null) {
-            if (other.element != null)
-                return false;
-        }
-        else if (!this.element.equals(other.element))
-            return false;
-        if (this.type == null) {
-            if (other.type != null)
-                return false;
-        }
-        else if (!this.type.equals(other.type))
-            return false;
-        return true;
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((this.code == null) ? 0 : this.code.hashCode());
-        result = prime * result + ((this.title == null) ? 0 : this.title.hashCode());
-        result = prime * result + ((this.level == null) ? 0 : this.level.hashCode());
-        result = prime * result + ((this.element == null) ? 0 : this.element.hashCode());
-        result = prime * result + ((this.type == null) ? 0 : this.type.hashCode());
-        return result;
-    }
-
-    @Override
-    public String toString() {
-        StringBuilder sb = new StringBuilder("CodeOfAccount (");
-
-        sb.append(code);
-        sb.append(", ").append(title);
-        sb.append(", ").append(level);
-        sb.append(", ").append(element);
-        sb.append(", ").append(type);
-
-        sb.append(")");
-        return sb.toString();
-    }
-}
