--- src/main/java/com/thee5176/record/springboot_cqrs_command/model/tables/records/CodeOfAccountRecord.java
@@ -1,147 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.thee5176.record.springboot_cqrs_command.model.tables.records;
-
-
-import com.thee5176.record.springboot_cqrs_command.model.enums.BalanceType;
-import com.thee5176.record.springboot_cqrs_command.model.enums.Element;
-import com.thee5176.record.springboot_cqrs_command.model.tables.CodeOfAccount;
-
-import org.jooq.Record1;
-import org.jooq.impl.UpdatableRecordImpl;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
-public class CodeOfAccountRecord extends UpdatableRecordImpl<CodeOfAccountRecord> {
-
-    private static final long serialVersionUID = 1L;
-
-    /**
-     * Setter for <code>public.code_of_account.code</code>.
-     */
-    public CodeOfAccountRecord setCode(Integer value) {
-        set(0, value);
-        return this;
-    }
-
-    /**
-     * Getter for <code>public.code_of_account.code</code>.
-     */
-    public Integer getCode() {
-        return (Integer) get(0);
-    }
-
-    /**
-     * Setter for <code>public.code_of_account.title</code>.
-     */
-    public CodeOfAccountRecord setTitle(String value) {
-        set(1, value);
-        return this;
-    }
-
-    /**
-     * Getter for <code>public.code_of_account.title</code>.
-     */
-    public String getTitle() {
-        return (String) get(1);
-    }
-
-    /**
-     * Setter for <code>public.code_of_account.level</code>.
-     */
-    public CodeOfAccountRecord setLevel(Integer value) {
-        set(2, value);
-        return this;
-    }
-
-    /**
-     * Getter for <code>public.code_of_account.level</code>.
-     */
-    public Integer getLevel() {
-        return (Integer) get(2);
-    }
-
-    /**
-     * Setter for <code>public.code_of_account.element</code>.
-     */
-    public CodeOfAccountRecord setElement(Element value) {
-        set(3, value);
-        return this;
-    }
-
-    /**
-     * Getter for <code>public.code_of_account.element</code>.
-     */
-    public Element getElement() {
-        return (Element) get(3);
-    }
-
-    /**
-     * Setter for <code>public.code_of_account.type</code>.
-     */
-    public CodeOfAccountRecord setType(BalanceType value) {
-        set(4, value);
-        return this;
-    }
-
-    /**
-     * Getter for <code>public.code_of_account.type</code>.
-     */
-    public BalanceType getType() {
-        return (BalanceType) get(4);
-    }
-
-    // -------------------------------------------------------------------------
-    // Primary key information
-    // -------------------------------------------------------------------------
-
-    @Override
-    public Record1<Integer> key() {
-        return (Record1) super.key();
-    }
-
-    // -------------------------------------------------------------------------
-    // Constructors
-    // -------------------------------------------------------------------------
-
-    /**
-     * Create a detached CodeOfAccountRecord
-     */
-    public CodeOfAccountRecord() {
-        super(CodeOfAccount.CODE_OF_ACCOUNT);
-    }
-
-    /**
-     * Create a detached, initialised CodeOfAccountRecord
-     */
-    public CodeOfAccountRecord(Integer code, String title, Integer level, Element element, BalanceType type) {
-        super(CodeOfAccount.CODE_OF_ACCOUNT);
-
-        setCode(code);
-        setTitle(title);
-        setLevel(level);
-        setElement(element);
-        setType(type);
-        resetChangedOnNotNull();
-    }
-
-    /**
-     * Create a detached, initialised CodeOfAccountRecord
-     */
-    public CodeOfAccountRecord(com.thee5176.record.springboot_cqrs_command.model.tables.pojos.CodeOfAccount value) {
-        super(CodeOfAccount.CODE_OF_ACCOUNT);
-
-        if (value != null) {
-            setCode(value.getCode());
-            setTitle(value.getTitle());
-            setLevel(value.getLevel());
-            setElement(value.getElement());
-            setType(value.getType());
-            resetChangedOnNotNull();
-        }
-    }
-}
